refactor : price, quantity 자료형 변경

+++ src/main/java/cart/dto/CartItemResponse.java
@@ -4,10 +4,10 @@ import cart.domain.CartItem;
 
 public class CartItemResponse {
     private Long id;
-    private int quantity;
+    private Long quantity;
     private ProductResponse product;
 
-    private CartItemResponse(Long id, int quantity, ProductResponse product) {
+    private CartItemResponse(Long id, Long quantity, ProductResponse product) {
         this.id = id;
         this.quantity = quantity;
         this.product = product;
@@ -25,7 +25,7 @@ public class CartItemResponse {
         return id;
     }
 
-    public int getQuantity() {
+    public Long getQuantity() {
         return quantity;
     }
